Make it compatible with GLIBCXX_3.4.19 (CentOS 7)
Fixes https://github.com/apache/pulsar-client-node/issues/286

### Motivation

The maximum GLIBCXX version on CentOS 7 is 3.4.19. If we want to use the
Pulsar Node client on CentOS 7, we have to upgrade the libstdc++.so. The
reason is the libstdc++ version is too high for some systems.
https://github.com/apache/pulsar-client-node/pull/285 downgrades the
base image to `node:18-buster` but it still requires the
`GLIBCXX_3.4.22`.

### Modifications
- Use `centos:7` as the base image to build `Pulsar.node` for
  glibc-based Linux distributions.
- Add `load_test.sh` to verify `Pulsar.node` can be loaded by other
  docker images and test Node.js 16 and 19 on Debian buster and
  bullseye.